gun polish

+++ Assets/Scripts/GunStats.cs
@@ -7,7 +7,7 @@ public class GunStats : MonoBehaviour
     public float damage;
     public float cooldown;
     public float range_or_duration;
-    public float max_bullet_count;
+    public int max_bullet_count;
     public float velocity;
     public float spread;
     public float size;
